+++ Engine/Source/GUI/Components/MenuComponent/Menu.cpp
@@ -13,17 +13,17 @@ MenuComponent(MenuComponent::height)
 void TextureGenEngine::Menu::Draw()
 {
         MenuComponent::Draw();
-    m_text->Draw(m_title, m_x, m_y, m_height, m_width, 12, glm::vec3(0.0f, 0.0f, 0.0f), AlignmentHorizontal::LEFT, AlignmentVertical::CENTER);
+        m_text->Draw(m_title, m_x, m_y, (int)m_height, (int)m_width, 12, glm::vec3(0.0f, 0.0f, 0.0f), AlignmentHorizontal::LEFT, AlignmentVertical::CENTER);
 }
 
-void TextureGenEngine::Menu::Resize(int width, int height)
+void TextureGenEngine::Menu::Resize(float width, float height)
 {
     m_y = height - m_height;
     m_background->SetPosition(m_x, m_y);
     m_width = width;
 }
 
-void TextureGenEngine::Menu::SetMenuBar(int x, int y, int width, int height)
+void TextureGenEngine::Menu::SetMenuBar(float x, float y, float width, float height)
 {
     m_x = x;
     m_y = y;
