+++ Engine/Source/GUI/Components/MenuComponent/Menu.h
@@ -16,8 +16,8 @@ namespace TextureGenEngine
     public:
         TAPI Menu(std::string title);
         void Draw() override;
-        void Resize(int width, int height) override;
-        void SetMenuBar(int x, int y, int width, int height);
+        void Resize(float width, float height) override;
+        void SetMenuBar(float x, float y, float width, float height);
         ~Menu();
     };
