+++ Engine/Source/GUI/Components/MenuComponent/MenuComponent.cpp
@@ -2,11 +2,11 @@
 #include "Core/Renderer/Mesh.h"
 #include "Core/Logger/Logger.h"
 
-TextureGenEngine::MenuComponent::MenuComponent(int height) : Component(0, 0, 100, height)
+TextureGenEngine::MenuComponent::MenuComponent(float height) : Component(0, 0, 100, height)
 {
 }
 
-void TextureGenEngine::MenuComponent::Init(int width, int height)
+void TextureGenEngine::MenuComponent::Init(float width, float height)
 {
     Component::Init(width, height);
     m_y = height - m_height;
@@ -19,7 +19,7 @@ void TextureGenEngine::MenuComponent::Draw()
     Component::Draw();
 }
 
-void TextureGenEngine::MenuComponent::Resize(int width, int height)
+void TextureGenEngine::MenuComponent::Resize(float width, float height)
 {
     m_y = height - m_height;
     m_background->SetPosition(m_x, m_y);
